Fix hover causing CSS glitches
When the on-hover list of items extends below the bottom of the page it triggers the global scrollbar, which slightly narrows the page, and may cause the hover to end if the mouse is positioned in just the wrong place. However, then the global scrollbar disappears and the hover gets triggered again, etc. This change adds a scrollbar within the list of items.